requirements.txt
openai>=1.0.0
anthropic>=0.18.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
plotly>=5.14.0
jupyter>=1.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.31.0
tiktoken>=0.5.0
transformers>=4.35.0
torch>=2.0.0
scikit-learn>=1.3.0
